Gun and Boss Basic implimentation
Gun is now able to actually be shot it will still play the animation but now you can't spam it as well as it shoots a line trace which can be set to detect specific items and objects.

A place holder boss is not set up and can even have it's panel shot off.

Basic widget with shitty dot and a health bar to show julian